compiler denoisification (msvc 18)

+++ src/lib/support/test/refcounted.cpp
@@ -117,7 +117,7 @@ BOOST_AUTO_TEST_CASE(test_support_refcounted_async)
     delete t;
   }
   
-  BOOST_CHECK(tpool.size() == a.get_ref());
+  BOOST_CHECK(tpool.size() == unsigned(a.get_ref()));
 
 #if 0
   BOOST_MESSAGE(a.get_ref());
